Move `list_experiments()` to `experiment.py`
